requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.7.4
PyJWT==2.8.0
python-multipart==0.0.9
